simplify TimeFieldsGroup usage
